requirements.txt
asn1crypto==0.24.0
bcrypt==3.1.4
boto3==1.5.18
botocore==1.8.32
cachetools==2.1.0
cffi==1.11.4
cryptography==2.1.4
docutils==0.14
enum34==1.1.6
Fabric==1.14.0
futures==3.2.0
google-api-python-client==1.6.7
google-auth==1.4.1
google-auth-httplib2==0.0.3
httplib2==0.11.3
idna==2.6
ipaddress==1.0.19
jmespath==0.9.3
marshmallow==3.0.0b8
oauth2client==4.1.2
paramiko==2.4.0
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyNaCl==1.2.1
python-dateutil==2.6.1
rsa==3.4.2
s3transfer==0.1.12
six==1.11.0
uritemplate==3.0.0
